Test `Array/windowed(in:allowsPartiality:)` and `Array/paired(to:)`
