+++ StandardKit/Chromodynamics/Array+Pair.swift
